Shrink @ts-nocheck surface: isolate the generateText call
The ai@7-beta type-inference OOM is triggered by the tool({inputSchema,execute})
+ generateText construction, so that whole call now lives in lib/run-agent.ts
behind a typed boundary (runAgent → { text, toolResults }). @ts-nocheck applies
only there; generate-response.ts (orchestration, failure-surfacing, tool-cache)
is now fully type-checked. Pure refactor — behavior unchanged.

Verified attempt to remove @ts-nocheck entirely (typing tools as ToolSet) still
OOMs tsc, so the directive stays on the isolated module until ai@7 stabilizes.
